Fix error in computation of SLS quasi-permanent combinations
The permanent loads alone (without any variable load) were missing
among those combinations. The error was sometimes masked by another
error in the code: for those quasi-permanent combinations the seismic
situation flag was incorrectly set to true. Due to this error, when all
the variable loads had a zero value for its psi_2 combination factor
the result was apparently correct.

+++ verif/tests/actions/load_combinations/ec0/test_ec0_accidental_combinations_02.py
@@ -45,5 +45,6 @@ if comparisonOK:
    print('test '+fname+': ok.')
 else:
     lmsg.error(fname+' ERROR.')
-os.remove(outputPath) # Your garbage, you clean it.
+if(comparisonOK):
+    os.remove(outputPath) # Your garbage, you clean it.
