+++ verif/tests/actions/load_combinations/ec0/test_ec0_seismic_combinations.py
@@ -41,4 +41,5 @@ if comparisonOK:
    print('test '+fname+': ok.')
 else:
     lmsg.error(fname+' ERROR.')
-os.remove(outputFileName) # Your garbage, you clean it.
+if comparisonOK:
+    os.remove(outputFileName) # Your garbage, you clean it.
